// src/lib/types/football.ts
export type TeamProfile = {
  id: string;
  name: string;
  fifaRanking: number;
  recent5Results: string[];
  recent10Results: string[];
  goalsForLast5: number;
  goalsAgainstLast5: number;
  xGLast5: number;
  xGALast5: number;
  lineupStability: number;
  manager: string;
  formation: string;
  pressingLevel: number;
  counterAttackAbility: number;
  possessionAbility: number;
  setPieceThreat: number;
  wingAttackRatio: number;
  centralProgression: number;
  vulnerableToCounter: number;
  vulnerableToPress: number;
  slowStarter: boolean;
  firstHalfGoalRate: number;
  secondHalfGoalRate: number;
  winRateAfterScoringFirst: number;
  comebackAbility: number;
  majorTournamentExperience: number;
  penaltyShootoutAbility: number;
  weatherAdaptability: number;
  regionAdaptability: number;
};

export type PlayerProfile = {
  id: string;
  name: string;
  age: number;
  nationalTeam: string;
  club: string;
  primaryClubPosition: string;
  nationalTeamPosition: string;
  isPositionMismatch: boolean;
  recentClubForm: number;
  recentNationalTeamForm: number;
  minutesLast10: number;
  goalsLast5: number;
  assistsLast5: number;
  shotsLast5: number;
  shotsOnTargetLast5: number;
  keyPassesLast5: number;
  tacklesLast5: number;
  interceptionsLast5: number;
  foulsLast5: number;
  fouledLast5: number;
  yellowCardsLast5: number;
  redCardsLast5: number;
  injuryStatus: "fit" | "minor" | "doubtful" | "out";
  injuryHistory: string[];
  recentlyReturnedFromInjury: boolean;
  fatigueRisk: number;
  travelRisk: number;
  climateFit: number;
  tacticalFit: number;
  positionFitScore?: number;
};

export type Match = {
  id: string;
  date: string;
  venue: string;
  stage: string;
  homeTeamId: string;
  awayTeamId: string;
  weather: string;
  refereeProfile: string;
  narrative: string;
  actualResult?: string;
};